+++ packages/cli/src/run.ts
@@ -1,6 +1,6 @@
 import pathParser from 'node:path'
 
-import { build, ParallelPluginError, PluginError } from '@kubb/core'
+import { build, ParallelPluginError, PluginError, timeout } from '@kubb/core'
 
 import { execa } from 'execa'
 import pc from 'picocolors'
@@ -48,6 +48,7 @@ export async function run({ config, options, spinner }: RunProps): Promise<void>
       const [cmd, ..._args] = [...parseArgsStringToArgv(command)]
       spinner.start(parseText(`🪂 Executing hooks(${pc.yellow('done')})`, { info: ` ${pc.dim(command)}` }, logLevel))
       const { stdout } = await execa(cmd, _args)
+      await timeout(200)
       spinner.succeed(parseText(`🪂 Executing hooks(${pc.yellow('done')})`, { info: ` ${pc.dim(command)}` }, logLevel))
 
       if (logLevel === 'info') {
@@ -69,8 +70,12 @@ export async function run({ config, options, spinner }: RunProps): Promise<void>
       if (!a.meta?.pluginName || !b.meta?.pluginName) {
         return 0
       }
-      if (a.meta?.pluginName.length < b.meta?.pluginName.length) return 1
-      if (a.meta?.pluginName.length > b.meta?.pluginName.length) return -1
+      if (a.meta?.pluginName.length < b.meta?.pluginName.length) {
+        return 1
+      }
+      if (a.meta?.pluginName.length > b.meta?.pluginName.length) {
+        return -1
+      }
       return 0
     })
 
@@ -99,12 +104,6 @@ ${pc.bold('Generated:')}      ${meta.filesCreated} files
 
   const printErrors = (error: Error) => {
     const pe = new PrettyError()
-
-    if (error instanceof ParallelPluginError) {
-      error.errors.map((e) => printErrors(e))
-
-      return
-    }
 
     if (options.debug) {
       spinner.fail(pc.red(`Something went wrong\n\n`))
@@ -150,7 +149,11 @@ ${pc.bold('Generated:')}      ${meta.filesCreated} files
 
     printSummary(output.pluginManager, 'success')
   } catch (error: any) {
-    printErrors(error as Error)
+    if (error instanceof ParallelPluginError) {
+      error.errors.map((e) => printErrors(e))
+    } else {
+      printErrors(error as Error)
+    }
 
     if (error instanceof PluginError || error instanceof ParallelPluginError) {
       printSummary(error.pluginManager, 'failed')