+++ packages/core/src/index.ts
@@ -3,7 +3,7 @@ import { build } from './build.ts'
 export * from './config.ts'
 export * from './build.ts'
 export * from './types.ts'
-export { CorePluginOptions, createPlugin, name } from './plugin.ts'
+export { CorePluginOptions, createPlugin, pluginName, pluginName as name } from './plugin.ts'
 
 export * from './utils/index.ts'
 export * from './managers/index.ts'