+++ packages/swagger-zodios/src/plugin.ts
@@ -11,7 +11,7 @@ import { OperationGenerator } from './generators/index.ts'
 import type { API as SwaggerApi } from '@kubb/swagger'
 import type { PluginOptions } from './types.ts'
 
-export const pluginName = 'swagger-zodios' as const
+export const pluginName: PluginOptions['name'] = 'swagger-zodios' as const
 
 // Register your plugin for maximum type safety
 declare module '@kubb/core' {